+++ tests/test_export_avro.py
@@ -5,7 +5,8 @@ from typer.testing import CliRunner
 
 from datacontract.cli import app
 from datacontract.export.avro_converter import to_avro_schema_json
-from datacontract.model.data_contract_specification import DataContractSpecification
+from datacontract.model.data_contract_specification import \
+    DataContractSpecification
 
 logging.basicConfig(level=logging.DEBUG, force=True)
 
@@ -19,6 +20,17 @@ def test_cli():
 def test_to_avro_schema():
     data_contract = DataContractSpecification.from_file("fixtures/avro/export/datacontract.yaml")
     with open("fixtures/avro/export/orders_with_datefields.avsc") as file:
+        expected_avro_schema = file.read()
+
+    model_name, model = next(iter(data_contract.models.items()))
+    result = to_avro_schema_json(model_name, model)
+
+    assert json.loads(result) == json.loads(expected_avro_schema)
+
+
+def test_to_avro_schema_with_logicalTypes():
+    data_contract = DataContractSpecification.from_file("fixtures/avro/export/datacontract_logicalType.yaml")
+    with open("fixtures/avro/export/datacontract_logicalType.avsc") as file:
         expected_avro_schema = file.read()
 
     model_name, model = next(iter(data_contract.models.items()))